feat(aither-core): extract the WPA/EAPOL core, so the fuzzers reach shipped code
#819's finding: fuzz_wpa.rs and fuzz_eapol.rs exercised `aither`, which the
kernel does not link. The kernel parses EAPOL in its own wifi.rs. The fuzzing
protected code that never runs on a device, and a green run read as coverage.

`aither-core` now holds the canonical EAPOL types/parser/encoder, PMK/PTK/MIC
derivation, and the 4-way handshake state machine. Both sides link it, so the
existing corpus reaches shipped code with NO fuzz-target edit -- the targets
import the outer crate, exactly as fuzz_gsm7/fuzz_pdu import klesis.
check-fuzz-target-coverage.sh confirms it: fuzz_eapol and fuzz_wpa moved from
GAP to ok, coverage 5/11 -> 8/11, and their EXEMPT rows are deleted because
that check fails on a stale exemption as well as on a new gap.

**The crypto backend swap is the load-bearing part, and it was gated on
vectors rather than on the new code's own tests.** The kernel's hand-rolled
SHA-1 engine, HMAC-SHA1, PBKDF2-HMAC-SHA1 and PRF-384 are deleted in favour of
audited RustCrypto. Both vector sets now live in aither-core and pass against
the new backend: RFC 2202 (three cases, including the long-key case), RFC 6070
c=1 and c=4096, IEEE 802.11i Annex J's PMK vector, and H.7.1's PRF-384 vector.
"New passes its own tests" would not have been evidence; these are the same
vectors the old implementations passed.

That swap also deletes the code carrying two filed defects (#837): the silent
32-byte PBKDF2 salt truncation and the unbounded prf_384 stack buffer. A
regression test pins the first. security.rs keeps only a thin sha1() wrapper,
used by ekphrasis.rs's RFC 6455 WebSocket handshake and unrelated to WPA.

Two real divergences resolved toward the kernel, which was the stricter side
on every point of overlap:

- aither's parser accepted any EAPOL key-descriptor byte; the kernel rejects
  non-RSN. The gate is now shared (test: rejects_non_rsn_key_descriptor_type).
- aither's encode_key_frame capped only the key_data_length FIELD while
  writing the full untruncated key_data -- the same class as the kernel's
  audit #282 finding 5, which aither had never fixed. Found during the merge,
  not looked for.

The handshake state machine (HandshakeState/WpaHandshake) is a lift, not a
merge: aither had no equivalent, only a replay counter with no key_info or
message-order awareness. It is the highest-value piece to move because it was
fuzzed nowhere and is #[expect(dead_code)] pending #129's hardware TX/RX --
doing this before #129 makes it RF-reachable is the right order.

generate_snonce() stays in the kernel: it is CSPRNG-facing, so aither-core
cannot link it, and it fails closed with CsprngNotSeeded before the nonce
reaches the handshake.

`#[non_exhaustive]` dropped from EapolKeyFrame and EapolFrame only. Its purpose
is protecting consumers outside the crate that cannot be enumerated; here both
consumers are in-repo, and keeping it would push ~27 construction sites through
builders for no safety gained. It stays on the enums.

All three lockfiles regenerated -- crates/thumos and fuzz/ are each their own
resolution root, and a missed one fails --locked with an error that reads as
unrelated breakage.

Closes #819
Refs #545
