+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 members = [
     "crates/aither",
+    "crates/aither-core",
     "crates/asphaleia",
     "crates/asphaleia-core",
     "crates/eidolon",
